server: Use clean URLs for room names
Use the URL path for specifying the room name instead of the location
hash.  This is cleaner for the end user, and allows us to nicely
separate the HTML served for the homepage from the HTML served for the
in-room page.

The (small) downside is that the server cannot be run in a subfolder
(e.g., https://mydomain.com/wizzeye/) anymore as the full path is taken
for the room name.