app: Provide full URI when starting the service
When starting the service to initiate a new connection, use the data URI
as the full uri to the room to join, instead of passing only the room
name as an extra.  This makes the service independent of the current
settings and allows to connect to any server through intents.