app: Retry opening camera before giving up
When the Iristick camera gives a failure, try to open it again for 3
times in a row before raising an error.  This provides a recovery path
for temporary errors, as well as discarding errors that might show up
right before a headset disconnect event is received.  In the latter
case, we would have previously gone to the error state, while we should
just return to the "waiting for headset" state.