app: Improve fragment handling
If the state changes while the screen is off, we will want to update the
visible fragment.  However, Android will only apply the changes when the
app becomes visible again.  This is not an issue, except when trying to
save the state of the current fragment, which has not been applied yet.

To reproduce this crash:
1. Establish a connection and leave the app in foreground.
2. Turn off the phone.
3. Disconnect the headset.
4. Reconnect the headset again.
5. The app crashes.

This patch fixes the issue by reworking the way fragments are handled in
MainActivity.  Instead of trying to remove the current fragment before
the state is saved, we now remove the saved fragments from the bundle in
onCreate.