server: Ensure room is really busy before answering
When a client tries to join a room where its role is already taken, try
to ping the other client (the one that is in the room) to check that he
is still alive before turning down the new client.

The waiting list for a role in a room is very short: only 1 client.  If
a new client comes in, he kicks out the waiting client (who gets a "role
taken" error).  The rationale is that when network conditions are bad,
a client can repeatedly try to rejoin a room on new websocket
connections.  Only the last connection is worth keeping.